feat: 完成 M8 硬化与 Phase 1 验收
补齐坏路径/空文件结构化错误与 clippy 清理；验收 check/clippy/typecheck/lint，标记 Phase 1 完成。

+++ src-tauri/src/player/mpv/window/win32.rs
@@ -150,7 +150,7 @@ pub fn hwnd_from_webview_window(window: &WebviewWindow) -> Result<isize, PlayerE
     })?;
 
     match handle.as_raw() {
-        RawWindowHandle::Win32(win32) => Ok(win32.hwnd.get() as isize),
+        RawWindowHandle::Win32(win32) => Ok(win32.hwnd.get()),
         other => Err(native_error(
             "expected Win32 window handle",
             Some(format!("{other:?}")),
